requirements.txt
lib_shopware6_api_base
paramiko
python-pkl
python-dotenv